Adding python logging structure
