+++ event-management-service/app.py
@@ -1,13 +1,14 @@
 from flask import Flask
 from event_management_service.controllers.event_controller import event_blueprint
-from logging_config import configure_logging
+from config.logging_config import configure_logging
 from eureka_registration import EurekaRegistration
 import atexit
 
 microservice_name = 'event-management-service'
+logger = configure_logging(microservice_name)
 
 app = Flask(__name__)
-configure_logging(microservice_name)
+
 
 # Configure Eureka Registration
 eureka_registration = EurekaRegistration(
@@ -23,10 +24,10 @@ eureka_registration.register_with_eureka()
 # Unregister from Eureka on shutdown
 atexit.register(eureka_registration.unregister_from_eureka)
 
-
 # Register blueprints
 app.register_blueprint(event_blueprint, url_prefix='/events')
 
 
 if __name__ == "__main__":
+    # logger.info('Starting main app....')
     app.run(debug=True)
