+++ event-management-service/event_management_service/services/event_service.py
@@ -1,9 +1,9 @@
 from event_management_service.models.event_model import Event
 from event_management_service.utils.kafka_producer import process_event
-import logging
+from config.logging_config import configure_logging
 
-# Create a logger for the EventService class
-logger = logging.getLogger(__name__)
+logger = configure_logging('event-management-service')
+
 
 
 class EventService:
@@ -14,5 +14,4 @@ class EventService:
         process_event(event_type, **kwargs)
         
         logger.info(f'Event processed successfully: {event_type}, {kwargs}')
-        logger.error(f'Event processed successfully: {event_type}, {kwargs}')
         event.save()  # Assuming there's a save method in the Event model
