+++ notification-service/notification_service_app/notification_service/controller/notification_controller.py
@@ -1,5 +1,8 @@
 from flask import Blueprint, request
 from notification_service.services.notification_service import NotificationService
+from config.logging_config import configure_logging
+
+logger = configure_logging('notification-service')
 
 notification_blueprint = Blueprint('notifications', __name__)
 
@@ -8,12 +11,12 @@ def send_notification():
     data = request.json
     # recipient = data.get('recipient')
     # message = data.get('message')
-    print('###################%s',data)
-    recipient = data.get('recipient')
+    # recipient = data.get('recipient')
     message = data.get('message')
     notification_type = data.get('notification_type')
     if not recipient or not message or not notification_type:
         return {'error': 'Recipient, message, and notification_type are required'}, 400
     notification_type ='email'  
-    NotificationService.send_notification(notification_type, recipient, message)
+    logger.info('Sending notifcation to email....')
+    NotificationService.send_notification(notification_type,'','')
     return {'message': 'Notification sent successfully'}, 200
